Align public tree with google3 ahead of re-enabling the Piper export
The google3 copy of these files carries changes this repo does not: the
safevalues hardening in theme.service and analytics.service, MediaType moved
out of artifact-tab into core/models/types, 1p test scaffolding in six spec
files, and an @license header on audio-processor.js. Most were dropped by the
revert in #481.

Adopting them here makes the two trees identical under the Copybara export
transforms, so re-enabling the Piper -> GitHub postsubmit is a no-op rather
than a surprise content push. Reconciling in content is also the right layer:
expressing the difference as transform rules instead would permanently block
the safevalues hardening from ever reaching this repo.

MediaType keeps a re-export from artifact-tab.component.ts, so imports from
that path are unaffected, and the enum values are unchanged.

Verified: ng build clean, and ng test 628 passing -- identical to the
pre-change baseline.
