+++ src/app/components/system-instruction-diff-dialog/system-instruction-diff-dialog.component.spec.ts
@@ -15,6 +15,8 @@
  * limitations under the License.
  */
 
+// 1p-ONLY-IMPORTS: import {beforeEach, describe, expect, it}
+import {initTestBed} from '../../testing/utils';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
@@ -34,6 +36,7 @@ describe('SystemInstructionDiffDialogComponent', () => {
   beforeEach(async () => {
     mockDialogRef = jasmine.createSpyObj('MatDialogRef', ['close']);
 
+    initTestBed();  // required for 1p compat
     await TestBed.configureTestingModule({
       imports: [SystemInstructionDiffDialogComponent, NoopAnimationsModule],
       providers: [
