+++ src/app/core/models/trace/ExperimentalSemconv.spec.ts
@@ -15,6 +15,8 @@
  * limitations under the License.
  */
 
+// 1p-ONLY-IMPORTS: import {beforeEach, describe, expect, it}
+import {initTestBed} from '../../../testing/utils';
 import {
   CompletionDetailsLogValidator,
   GEN_AI_COMPLETION_DETAILS_EVENT,
