+++ src/app/core/services/analytics.service.ts
@@ -17,6 +17,8 @@
 
 import { inject, Injectable, effect } from '@angular/core';
 import { TelemetryService } from './telemetry.service';
+import {setScriptSrc} from 'safevalues/dom';
+import {trustedResourceUrl} from 'safevalues';
 
 declare global {
   interface Window {
@@ -66,7 +68,7 @@ export class AnalyticsService {
       // Load GA4 script
       const script = document.createElement('script');
       script.async = true;
-      script.src = `https://www.googletagmanager.com/gtag/js?id=${this.measurementId}`;
+      setScriptSrc(script, trustedResourceUrl`https://www.googletagmanager.com/gtag/js?id=${this.measurementId}`);
       document.head.appendChild(script);
 
       window.gtag('js', new Date());
